add flow filter component
