+++ src/graphql/Calendar/calendar-day-flow-runs.gql
@@ -1,0 +1,20 @@
+query CalendarDayFlowRuns($startTime: timestamptz, $endTime: timestamptz, $flowIds: [uuid!]) {
+  flow_run (
+    where: { 
+      _and: [
+    {flow_id: { _in: $flowIds}},
+    {start_time: { _gt: $startTime}}, 
+    {start_time: { _lt: $endTime}}
+      ]
+    }
+  ) 
+  {
+    id
+    flow_id
+    name
+    end_time
+    start_time
+    scheduled_start_time
+    state
+  }
+}
