+++ src/graphql/Calendar/calendar-day-ongoing-flow-runs.gql
@@ -1,0 +1,20 @@
+query CalendarDayOngoingFlowRuns($flowIds: [uuid!], $startTime:timestamptz, $endTime: timestamptz) {
+  flow_run (
+    where: { 
+      _and: [
+    {flow_id: { _in: $flowIds}},
+    {start_time: {_lt: $startTime}}
+    {end_time: {_gt: $endTime}}
+      ]
+    }
+  ) 
+  {
+    id
+    flow_id
+    name
+    end_time
+    start_time
+    scheduled_start_time
+    state
+  }
+}
