+++ src/graphql/Calendar/calendar-day-scheduled-flow-runs.gql
@@ -1,0 +1,21 @@
+query CalendarDayScheduledFlowRuns($project_id: uuid, $startTime: timestamptz, $endTime: timestamptz, $flowIds: [uuid!]) {
+  flow_run (
+    where: { 
+      _and: [
+    {scheduled_start_time: { _gt: $startTime}}, 
+    {scheduled_start_time: { _lt: $endTime}},
+    {start_time: {_is_null: true}},
+    {flow_id: { _in: $flowIds}}
+    ]
+    }
+  ) 
+  {
+    id
+    flow_id
+    name
+    end_time
+    start_time
+    scheduled_start_time
+    state
+  }
+}
