+++ src/graphql/Calendar/calendar-flow-runs.gql
@@ -1,21 +1,16 @@
-query CalendarFlowRuns($project_id: uuid, $startTime: timestamptz, $endTime: timestamptz, $flowIds: [uuid!]) {
+query CalendarFlowRuns($startTime: timestamptz, $endTime: timestamptz) {
   flow_run (
     where: { 
       _and: [
-    {flow: { project_id: { _eq: $project_id } }}
-    {flow_id: { _in: $flowIds}},
     {start_time: { _gt: $startTime}}, 
     {start_time: { _lt: $endTime}}
       ]
     }
+    distinct_on: flow_id
+    limit: 1
   ) 
   {
     id
     flow_id
-    name
-    end_time
-    start_time
-    scheduled_start_time
-    state
   }
 }
