+++ src/graphql/Calendar/calendar-running-flow-runs.gql
@@ -1,20 +1,16 @@
-query CalendarRunningFlowRuns($flowIds: [uuid!], $startTime:timestamptz) {
+query CalendarRunningFlowRuns($startTime:timestamptz) {
   flow_run (
     where: { 
       _and: [
-    {flow_id: { _in: $flowIds}},
     {state: { _eq: "Running"}},
     {start_time: {_lt: $startTime}}
       ]
     }
+    limit: 1
+    distinct_on: flow_id
   ) 
   {
     id
     flow_id
-    name
-    end_time
-    start_time
-    scheduled_start_time
-    state
   }
 }
