+++ src/graphql/Calendar/calendar-scheduled-flow-runs.gql
@@ -1,22 +1,15 @@
-query CalendarScheduledFlowRuns($project_id: uuid, $startTime: timestamptz, $endTime: timestamptz, $flowIds: [uuid!]) {
+query CalendarScheduledFlowRuns($startTime: timestamptz, $endTime: timestamptz) {
   flow_run (
     where: { 
       _and: [
-    {flow: { project_id: { _eq: $project_id } }}
     {scheduled_start_time: { _gt: $startTime}}, 
     {scheduled_start_time: { _lt: $endTime}},
     {start_time: {_is_null: true}},
-    {flow_id: { _in: $flowIds}}
     ]
     }
   ) 
   {
     id
     flow_id
-    name
-    end_time
-    start_time
-    scheduled_start_time
-    state
   }
 }
